split models from modules
