+++ internal/app/model/release.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT
 // license that can be found in the LICENSE file.
 
-package module
+package model
 
 import (
 	"encoding/json"
