+++ internal/app/model/release_path.go
@@ -1,0 +1,12 @@
+// Copyright 2019 Clivern. All rights reserved.
+// Use of this source code is governed by the MIT
+// license that can be found in the LICENSE file.
+
+package model
+
+// ReleasePath struct
+type ReleasePath struct {
+	VCS        string
+	Author     string
+	Repository string
+}
